requirements.txt
Django==2.0.1
djangorestframework==3.7.7
markdown==2.6.11
psycopg2-binary==2.7.6.1
django-filter==2.4.0
django-debug-toolbar==1.9.1
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
django-extensions==1.9.9
django-rest-auth==0.9.2
django-jet==1.0.7
django-otp==0.5.2
qrcode==5.3
django-versatileimagefield==1.8
mammoth==1.4.3
django-haystack==2.7.dev1
pysolr==3.7.0
drf-haystack==1.7.1rc2
django-haystack-panel==0.2.1
sorl-thumbnail==12.4.1
moviepy==0.2.3.2
django-widget-tweaks==1.4.1
celery==4.1.0
WeasyPrint==0.42.3
django-colorfield==0.1.14
django-widget-tweaks==1.4.1
django-formtags==0.1.2
fcm-django==0.2.19
kombu==4.1.0
wordcloud==1.5.0
matplotlib==2.2.3
redis==2.10.6
django-celerybeat-status==0.0.10
django-multiselectfield==0.1.8
